[proliferate] move agents domain under domains
