+++ anyharness/crates/anyharness-lib/src/acp/session_actor.rs
@@ -21,8 +21,8 @@ use super::permission_broker::{
 use super::persistence_sanitizer::sanitize_raw_notification_for_sqlite;
 use super::provider_errors::{classify_provider_rate_limit_error, PROVIDER_RATE_LIMIT_CODE};
 use super::runtime_client::RuntimeClient;
-use crate::agents::model::{AgentKind, ResolvedAgent};
 use crate::api::http::latency::{latency_trace_fields, LatencyRequestContext};
+use crate::domains::agents::model::{AgentKind, ResolvedAgent};
 use crate::plans::model::{NewPlan, PlanRecord};
 use crate::plans::service::{PlanCreateError, PlanDecisionError, PlanService};
 use crate::reviews::service::ReviewService;
@@ -5057,11 +5057,11 @@ mod tests {
     use crate::acp::background_work::{BackgroundWorkOptions, BackgroundWorkRegistry};
     use crate::acp::event_sink::{SessionEventSink, SessionEventSinkDebugSnapshot};
     use crate::acp::permission_broker::{InteractionBroker, PermissionOutcome};
-    use crate::agents::model::{
+    use crate::domains::agents::model::{
         AgentKind, ArtifactRole, CredentialState, ResolvedAgent, ResolvedAgentStatus,
         ResolvedArtifact,
     };
-    use crate::agents::registry::built_in_registry;
+    use crate::domains::agents::registry::built_in_registry;
     use crate::persistence::Db;
     use crate::plans::{service::PlanService, store::PlanStore};
     use crate::sessions::live_config::{snapshot_to_record, NormalizedControlKind};
