+++ anyharness/crates/anyharness-lib/src/api/http/provider_configs.rs
@@ -8,12 +8,12 @@ use anyharness_contract::v1::{
 };
 use axum::{extract::State, Json};
 
-use crate::agents::model::{
+use crate::app::AppState;
+use crate::domains::agents::model::{
     ModelCatalogStatus, ModelLaunchRemediationKind, ModelLaunchRemediationMetadata,
     ModelRegistryMetadata, ModelRegistryModelMetadata, SessionDefaultControlKey,
     SessionDefaultControlMetadata, SessionDefaultControlValueMetadata,
 };
-use crate::app::AppState;
 
 #[utoipa::path(
     get,
