+++ anyharness/crates/anyharness-lib/src/app/mod.rs
@@ -4,15 +4,15 @@ use std::sync::Arc;
 use crate::acp::manager::AcpManager;
 use crate::adapters::git::WorkspaceFileSearchCache;
 use crate::adapters::processes::ProcessService;
-use crate::agents::catalog::{LaunchCatalogService, ModelCatalogService};
-use crate::agents::reconcile_execution::AgentReconcileService;
-use crate::agents::seed::AgentSeedStore;
 use crate::cowork::artifacts::CoworkArtifactRuntime;
 use crate::cowork::delegation::service::CoworkDelegationService;
 use crate::cowork::mcp_auth::CoworkMcpAuth;
 use crate::cowork::runtime::{CoworkRuntime, CoworkSessionHooks};
 use crate::cowork::service::CoworkService;
 use crate::cowork::store::CoworkStore;
+use crate::domains::agents::catalog::{LaunchCatalogService, ModelCatalogService};
+use crate::domains::agents::reconcile_execution::AgentReconcileService;
+use crate::domains::agents::seed::AgentSeedStore;
 use crate::mobility::service::MobilityService;
 use crate::mobility::store::MobilityStore;
 use crate::persistence::Db;
