+++ anyharness/crates/anyharness-lib/src/domains/agents/credentials.rs
