+++ anyharness/crates/anyharness-lib/src/domains/agents/install_lock.rs
