+++ anyharness/crates/anyharness-lib/src/domains/agents/registry.rs
